hack/setup_external_dev_services: Improve postgresql installation
The postgresql helm chart requires the cluster to provide a PersistentVolume
and places a PersisentVolumeClaim for this.
Ensure that the cluster provides PersistentVolumes.

Also set volumePermissions.enabled=true which helps to fix Permission denied
when trying to create the data directory in the PV mount directory
if the PV's permissions are not open enough:
https://github.com/bitnami/charts/issues/1210#issuecomment-1145923630