feature(peridot-cli): add `task logs` command
`peridot task logs` will accept either a build ID or package name. In
the latter case, the project will be queried for the latest build for
the given package name. With the build task ID in hand, the task can be
queried for its BUILD_ARCH subtasks, which contain the logs for the
binary RPM builds.

Passing `-C` or `--cwd` with a directory will change into that directory
before writing files. That directory **must** exist beforehand.
Otherwise, $PWD is used.

By default, a separate file will be created for each
<task>-<subtask>-<architecture> tuple, which accounts for modular packages
with multiple outputs.

Passing the `-c` or `--combined` flag will combine all the logs into a
single file, named <task id>.log